requirements.txt
Django>=4.2
channels>=4.0
djangorestframework>=3.14
channels-redis>=4.0
psycopg2-binary>=2.9
websockets>=12.0
pytest>=8.0
pytest-django>=4.5
pytest-asyncio>=0.23
daphne>=4.0
